requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
Django==4.2.2
djangorestframework==3.14.0
executing==1.2.0
idna==3.4
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyproj==3.6.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
urllib3==2.0.3
wcwidth==0.2.6
